Moved wsaddress msw into test
